requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# SMTP Server
aiosmtpd>=1.4.4
aiofiles>=23.1.0

# Email processing
# python-email-validator>=2.0.0
email-validator>=2.0.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# QKD and Quantum Security
requests>=2.31.0          # For Key Manager REST API
pycryptodome>=3.19.0      # Advanced cryptographic operations
qiskit>=0.45.0            # Quantum computing framework (optional)
numpy>=1.24.0             # Mathematical operations for quantum crypto

# Database
psycopg2-binary>=2.9.0  # PostgreSQL
mysqlclient>=2.2.0      # MySQL (optional)

# Caching and Performance
redis>=4.5.0
celery>=5.3.0
django-redis>=5.3.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0
ipython>=8.14.0

# Production
gunicorn>=21.2.0
uvicorn>=0.23.0
whitenoise>=6.5.0

# Monitoring and Logging
sentry-sdk>=1.29.0
structlog>=23.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Utilities
python-decouple>=3.8
python-dotenv>=1.0.0
pillow>=10.0.0

Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
aiosmtpd==1.4.4
dnspython==2.4.2
dkimpy==1.1.5
# spf==2.0.14
cryptography==41.0.7
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
email-validator==2.1.0
python-magic==0.4.27
requests==2.31.0
pycryptodome==3.19.0
